### 2025-07-31 #8683 @editorialbot
[PRE REVIEW]: OptiGob: A decision support tool for Agriculture, Forestry and Other Land Use transitions.
**Submitting author:** @colmduff (<a href="http://orcid.org/0000-0002-0076-6749">Colm Duffy</a>)
**Repository:** https://github.com/colmduff/OptiGob
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.1.3
