### 2025-08-09 af31927
paper update
