### 2025-08-27 #8683 @colmduff
> [@colmduff](https://github.com/colmduff)
>
> * I see a docs directory but no link to the docs. Can they be hosted on readthedocs or similar?
> * I saw your note about a tool that hasn't been released but I didn't follow it. Can you help explain?

Hi @kthyng

The documentation is just hosted on github pages at the moment https://colmduff.github.io/OptiGob/html/index.html

I can host it else where if that is necessary?

The package is the backend for this dashboard: http://goblin.cs.universityofgalway.ie/

I just thought it might be useful to see what it actually does.
