### 2025-08-27 #8683 @colmduff
> [@colmduff](https://github.com/colmduff)
>
> * also can you shorten your paper closer to 1000 words

@kthyng

if you can give me a ceiling that you would be happy with? Its about 1,500 at the moment. I can of course make every effort to reduce, but I just want to know if there is some flexibility on the 1000 or is that a hard limit.

Just a quick fyi, I am working to a deadline at the moment, so it may be a week or so before I have the scope to engage on this element.
