### 2025-08-29 #8683 @colmduff
> [@colmduff](https://github.com/colmduff)
>
> > The documentation is just hosted on github pages at the moment https://colmduff.github.io/OptiGob/html/index.html
>
> Perfect. I don't see this linked in your readme though?
>
> > if you can give me a ceiling that you would be happy with? Its about 1,500 at the moment. I can of course make every effort to reduce, but I just want to know if there is some flexibility on the 1000 or is that a hard limit.
>
> Under 1000 is best, but if you need a bit more finish a useful paragraph or section, that is probably ok. Otherwise, move it to the docs!

@kthyng , thanks a million.

I have added the doc link to the about section on the repo, and I have added a docs online badge in the readme.

working on word count now.
