### 2025-08-29 fb1f111
word count for paper and readme update
