+++ joss_paper/paper/paper.md
@@ -31,44 +31,35 @@ bibliography: paper.bib
 ---
 # Summary
 
-`OptiGob` is a Python-based tool designed to explore configurations of Ireland’s agriculture, forestry, and other land use (AFOLU) sectors out to 2100. The main purpose is to assist users in the assessment environmental and economic impact pathways, based on different land use transition pathways under varying assumptions about agricultural, afforestation, emissions abatement, and carbon dioxide removal (CDR) strategies. OptiGOB combines data outputs from the GOBLIN Lite for agriculture and land use, from FERS-CBM for forestry, and from the LCAD2.0 for anaerobic digestion, to generate scenarios that respect biophysical constraints. `OptiGob` provides a flexible, customizable, tool for researchers, policymakers, and educators to explore environmental and economic trade-offs associated with land use transition pathways.
-
+`OptiGob` is a Python-based tool designed to explore configurations of Ireland’s agriculture, forestry, and other land use (AFOLU) sectors. The main purpose is to assist users in the assessment of environmental and economic impact pathways, based on different land use transition pathways under varying assumptions about agricultural, afforestation, emissions abatement, and carbon dioxide removal (CDR) strategies. OptiGOB combines data outputs from the GOBLIN Lite for agriculture and land use, from FERS-CBM for forestry, and from the LCAD2.0 for anaerobic digestion, to generate scenarios that respect biophysical constraints. `OptiGob` provides a flexible, customisable, tool for researchers, policymakers, and educators to explore environmental and economic trade-offs associated with land use transition pathways.
 
 # Statement of need
 
-Ireland's agricultural landscape is dominated by grassland, supporting extensive dairy and suckler beef production. Investment in Ireland's bioeconomy requires substantial changes to Agriculture, Forestry and Other Land Use (AFOLU) sector. `OptiGob` allows users to manipulate critical levers within the AFOLU sector that determine transition pathways and systematically explore their impacts. `OptiGob` utilises data generated by `GOBLIN Lite` [@Duffy:2024; @DuffyB:2022], FERS-CBM [@Black:2025; @Kurz:2008], and LCAD2.0 [@Martinez-Acre:2025] to calculate outputs from agricultural production, forestry, and anaerobic digestion, respectively.
+Ireland's agricultural landscape is dominated by grassland, supporting extensive dairy and suckler beef production. Investment in Ireland's bio-economy requires substantial changes to AFOLU sector. `OptiGob` allows users to manipulate critical AFOLU levers that determine transition pathways to explore their impacts. `OptiGob` utilises data from `GOBLIN Lite` [@Duffy:2024; @DuffyB:2022], FERS-CBM [@Black:2025; @Kurz:2008], and LCAD2.0 [@Martinez-Acre:2025] to calculate outputs from agriculture, forestry, and anaerobic digestion.
 
-`OptiGob` model estimates the net greenhouse gas (GHG) emissions from land use change, based on user input, and estimates available emissions and land budget for grass-based livestock (Dairy and Suckler cow) production. Pyomo [@bynum2021pyomo; @hart2011pyomo] is used to optimise livestock populations, while respecting area commitment (afforestation, anaerobic digestion, BECCS, protein crops) and emission (CO~2~e in the case of Net-Zero, and CH~4~, along with Net-Zero N~2~O and CO~2~ in the case of split gas targets) constraints.
[…]
