+++ tests/data/sip.yaml
@@ -1,7 +1,7 @@
 AR: 5
 split_gas: true
 split_gas_frac: 0.3
-target_year: 2050
+target_year: 2030
 abatement_type: "frontier"
 abatement_scenario: 9
 livestock_ratio_type: "dairy_per_beef"
