### 2025-09-03 #8683 @colmduff
kanishkan91, varsha2509, thinkswhat, nicrie, Sierra-MC, david-yannick.

Of those listed above, I am fairly sure that varsha and david-yannick had reviewed a paper previously.

Not sure if this is helpful or not, but I use the package in an online dashboard. Perhaps reviewers will find it useful to see the outputs: https://goblin.cs.universityofgalway.ie/
