### 2025-09-16 #8683 @mengqi-z
👋 @varsha2509 @david-yannick - Would any of you be interested in reviewing this submission for JOSS? Since both of you have reviewed the GOBLIN Lite paper, you already have relevant knowledge, as this software builds on GOBLIN outputs to conduct land use change and environmental assessment.

JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
