### 2025-10-08 #8683 @mengqi-z
👋 @tscheypidi @iosefa @matthiasdemuzere  - Would any of you be interested in and available to review this submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
