### 2025-10-17 #9271 @mengqi-z
👋 @quaz115 @lionel42 - This is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

These checklists contain the JOSS requirements. As you go over the submission, **please check any items that you feel have been satisfied**. The first comment in this thread also contains links to the JOSS reviewer guidelines.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, **please mention `openjournals/joss-reviews#9271` so that a link is created to this thread** (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about **4-6 weeks**. Please let me know if any of you require some more time. We can also use EditorialBot (`@editorialbot`) to set automatic reminders if you know you'll be away for a known period of time.
