### 2025-10-22 colmduff/OptiGob#9 @lionel42
Source files of the doc on the repo
I noticed that in docs/html you have all the generated html of the doc. It is a common practice to not include these files since they are autogenerated.
Would it be okay to remove them ?

openjournals/joss-reviews#9271
