### 2025-10-27 colmduff/OptiGob#10 @lionel42
Lot of duplicates code
Hi,
As part of the joss review: openjournals/joss-reviews#9271

I noticed that the code contains files with a lot of duplicate code. This is in general not good for maintablity or extendability.

As examples:

These seems to functions to access values from the sql database. It would be better to have a single function that can specify which variable to access.
https://github.com/colmduff/OptiGob/blob/main/src/optigob/resource_manager/optigob_data_manager.py
https://github.com/colmduff/OptiGob/blob/main/src/optigob/resource_manager/database_manager.py

In most other folders, you have two files, one for baseline and one for the budget These files are almost duplicates. https://github.com/colmduff/OptiGob/blob/main/src/optigob/livestock/baseline_livestock.py and https://github.com/colmduff/OptiGob/blob/main/src/optigob/livestock/livestock_budget.py

This seems to be the case for most of the sectors you have.

If you want to make this package reusable and extendable, you should refactor the code to avoid all these duplicates.
