+++ docs/html/_sources/autoapi/optigob/index.rst.txt
@@ -1,39 +1,0 @@
-optigob
-=======
-
-.. py:module:: optigob
-
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 1
-
-   /autoapi/optigob/bioenergy/index
-   /autoapi/optigob/budget_model/index
-   /autoapi/optigob/database/index
-   /autoapi/optigob/forest/index
-   /autoapi/optigob/input_helper/index
-   /autoapi/optigob/livestock/index
-   /autoapi/optigob/optigob/index
-   /autoapi/optigob/other_land/index
-   /autoapi/optigob/protein_crops/index
-   /autoapi/optigob/resource_manager/index
-   /autoapi/optigob/static_ag/index
-   /autoapi/optigob/substitution/index
-
-
-Attributes
-----------
-
-.. autoapisummary::
-
-   optigob.__version__
-
-
-Package Contents
-----------------
-
-.. py:data:: __version__
-
