+++ docs/html/_sources/autoapi/optigob/budget_model/emissions_budget/index.rst.txt
@@ -1,450 +1,0 @@
-optigob.budget_model.emissions_budget
-=====================================
-
-.. py:module:: optigob.budget_model.emissions_budget
-
-.. autoapi-nested-parse::
-
-   Emissions Budget Module
-   =======================
-
-   This module defines the EmissionsBudget class, which calculates and aggregates emissions budgets and categories
-   across all sectors using the optigob_data_manager. Emissions are returned in kilotons (kt).
-
-   Class:
-       EmissionsBudget: Calculates total and sectoral emissions (CO2e, CO2, CH4, N2O), net zero and split gas budgets, and substitution emissions.
-
-   .. method:: __init__(self, optigob_data_manager)
-
-      
-      Initializes the EmissionsBudget with the provided data manager and sets up all sectoral budget classes and emission methods.
-
-   .. method:: _get_total_beccs_co2e(self)
-
-      
-      Calculates total BECCS CO2e emissions (kt).
-
-   .. method:: _get_total_beccs_co2(self)
-
-      
-      Calculates total BECCS CO2 emissions (kt).
-
-   .. method:: _get_total_beccs_ch4(self)
-
-      
-      Calculates total BECCS CH4 emissions (kt).
-
-   .. method:: _get_total_beccs_n2o(self)
-
-      
-      Calculates total BECCS N2O emissions (kt).
-
-   .. method:: _get_total_emission_co2e_budget(self)
-
-      
-      Calculates the total CO2e emissions budget (kt) for net zero.
-
-   .. method:: _get_total_emission_co2e(self)
-
-      
-      Calculates the current total CO2e emissions (kt).
-
-   .. method:: _split_gas_emissions_total_budget_co2e(self)
-
-      
-      Calculates the total split gas emissions budget (kt).
-
-   .. method:: _get_total_emission_ch4(self)
-
-      
-      Calculates the current total CH4 emissions (kt).
-
-   .. method:: _get_total_emission_n2o(self)
-
-      
-      Calculates the current total N2O emissions (kt).
-
-   .. method:: _get_total_emission_co2(self)
-
-      
-      Calculates the current total CO2 emissions (kt).
-
-   .. method:: _check_net_zero_status(self)
-
-      
-      Checks if the net zero budget is met (returns True/False).
-
-   .. method:: check_status(self)
-
-      
-      Returns a dict with the status of net zero and split gas budgets.
-
-   .. method:: _check_split_gas_net_zero_status(self)
-
-      
-      Checks if the split gas budget is met (returns True/False).
-
-   .. method:: get_split_gas_budget(self)
-
-      
-      Returns the split gas budget (kt).
-
-   .. method:: get_net_zero_budget(self)
-
-      
-      Returns the net zero budget (kt).
-
-   .. method:: total_agriculture_co2e_emission(self)
-
-      
-      Calculates total agriculture CO2e emissions (kt).
-
-   .. method:: total_agriculture_co2_emission(self)
-
-      
-      Calculates total agriculture CO2 emissions (kt).
-
-   .. method:: get_total_agriculture_ch4_emission(self)
-
-      
-      Calculates total agriculture CH4 emissions (kt).
-
-   .. method:: get_total_agriculture_n2o_emission(self)
-
-      
-      Calculates total agriculture N2O emissions (kt).
-
-   .. method:: get_co2e_emission_categories(self)
-
-      
-      Returns CO2e emissions by sector/category (kt).
-
-   .. method:: get_co2_emission_categories(self)
-
-      
[…]
