+++ docs/html/_sources/autoapi/optigob/protein_crops/index.rst.txt
@@ -1,15 +1,0 @@
-optigob.protein_crops
-=====================
-
-.. py:module:: optigob.protein_crops
-
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 1
-
-   /autoapi/optigob/protein_crops/protein_crops_budget/index
-
-

[…]
