### 2025-11-02 31db95c
removed joss folder from main
