+++ joss_paper/utils/config.py
@@ -1,5 +1,0 @@
-DEFAULT_FIGSIZE = (8, 6)
-DEFAULT_CMAP = "tab20"
-DEFAULT_SCENARIO_LABELS = ["Baseline", "Scenario"]
-INPUT_DATA_PATH = "../data/sip.yaml"
-FIGURES_PATH = "../figures"
