+++ joss_paper/data/sip.yaml
@@ -1,23 +1,0 @@
-AR: 5
-split_gas: true
-split_gas_frac: 0.3
-target_year: 2050
-abatement_type: "frontier"
-abatement_scenario: 9
-livestock_ratio_type: "dairy_per_beef"
-livestock_ratio_value: 10
-forest_harvest_intensity: low
-afforestation_rate_kha_per_year: 16
-broadleaf_fraction: 0.3
-organic_soil_fraction: 0
-beccs_included: true
-beccs_willow_area_multiplier: 1.5
-wetland_restored_frac: 0.9
-organic_soil_under_grass_frac: 0.5
-biomethane_included: true
-protein_crop_included: true
-protein_crop_multiplier: 0
-pig_and_poultry_multiplier: 1.2
-baseline_year: 2020
-baseline_dairy_pop: 156.8
-baseline_beef_pop: 98.4
