### 2025-11-02 ceb871b
updated focs and readthedocs file
