+++ .readthedocs.yml
@@ -15,7 +15,7 @@ build:
 sphinx:
   configuration: docs/conf.py
 
-# Optionally declare the Python requirements required to build your docs
+# Only install documentation requirements (not the full package)
 python:
   install:
     - requirements: docs/requirements.txt
