### 2025-11-02 6aa2afa
example output debug
