### 2025-11-02 364a3cf
example output debug
