### 2025-11-03 colmduff/OptiGob#9 @lionel42
Thanks a lot,
Looks very good.

please also adapt this link on the repo:
