### 2025-11-03 colmduff/OptiGob#19 @lionel42
Adding input validation
If you specify wrong set of parameters in the config.yml, you will get this error:

```
ValueError: No matching scaler found for the provided parameters.
```

For example with afforestation rates, it is raised at run time, when optigob tries to read the data:

```
  File "OptiGob/src/optigob/resource_manager/optigob_data_manager.py", line 522, in get_forest_scaler
    raise ValueError("No matching scaler found for the provided parameters.")
```

since you anyway read the data to extract all the inputs available like this:

```
    helper = InputHelper()

    helper.print_readable_combos(12)
```

which gives:

```
Input type: forest
  affor_rate_kha-yr | broadleaf_frac | organic_soil_frac | forest_harvest_intensity
  -----------------------------------------------------------------------------------
  2 | 0.5 | 0.15 | high
  8 | 0.5 | 0.15 | high
  8 | 0.5 | 0.15 | low
  8 | 0.3 | 0.15 | high

...
```

What about raising a more formal error about which parameters are not valid ?

https://github.com/openjournals/joss-reviews/issues/9271
