### 2025-11-03 colmduff/OptiGob#20 @lionel42
Joss paper review
I have finished looking at the code. Here is my review of the paper itself.

https://github.com/openjournals/joss-reviews/issues/9271

Lines and comments refer to this version : https://raw.githubusercontent.com/openjournals/joss-papers/joss.09271/joss.09271/10.21105.joss.09271.pdf

* lines 12-13, in other places in the paper you use the special software police for `optigob` and `GOBLIN Lite`, but here not. This would also apply to Pyomo that is reference below ?
* line 27, you introduce BECCS initials only at line 44. You should move it up here as you refer to it for the first time.
* line 28. `CO2e in the case of Net-Zero, and CH4, along with Net-Zero N2O and CO2 `, did you mean `CH4 along with CO2e, N2O and CO2` ?
* line 35: add YAML input files
* line 47: define GWP100
* Model Overview section: I would also mention the sqllite database, how it is generated and add it on Figure 1
* Figure 1: text is a bit too small, would be nice that it fits to the text size of the paper
* Figure 1: It is not clear what the arrows represent (data flow, execution order, dependency ?)
* paragraph line40 -> 44, this is very detailed would it not be kept out of the paper ?
* paragraph line 45->49: this is the execution order of the software,  I suggest a bullet point structure to improve readability.
* line 48: you already cited pyomo above, one time should be enough
* paragraph 50->53: bullet points here again ?
* paragraph 62->67: I would leave the details of your example parameters outside the paper and simply refer to an example in your doc or repo
* Section Example output: As a reviewer I am suppose to reproduce the results in the paper. Any chance  I can run the code for that ?  I have seen this file: https://github.com/colmduff/OptiGob/blob/joss/joss_paper/notebook/01_joss_generation.py but I am not sure how I am supposed to run it ;)
* Section Functionality: This seems to duplicate most of what was said in model overview. I would remove it and move some parts you find important to model overview. (like saying you use linear optimization)
* Section Limitations: It is very good to point out the limitations of your software, but  I am not sure that the paper is the right place to put all the limitations. They would fit better in the documentation (especially if you improve optigob in the future, you don't want these hard-coded in your paper)
* line 88: livestock instead of animal ?
* State of the field: You do not describe much of  "how this software compares to other commonly-used packages?" . Can you add this in statement of need if you now any similar software ?
