### 2025-11-03 colmduff/OptiGob#21 @lionel42
Improve the contributing section of the doc
It seems that you have just copied it from a template: https://optigob.readthedocs.io/en/latest/contributing.html#contributing

https://github.com/openjournals/joss-reviews/issues/9271

for joss standard you need:
```
Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
```

Add at least a link to the repository issues page.

Improving the text to optigob would also be nice.
