### 2025-11-05 #9271 @lionel42
Thanks for the explanations @colmduff

>
> Regarding the contributors: the other authors have not contributed code directly to the repository, which is why they don’t appear in the GitHub contributors list. However, they have been heavily involved in the development and curation of the underlying scenario database and model assumptions, work that represents a substantial part of the project. I’ll be happy to clarify these contributions in the paper and metadata as needed.

@mengqi-z  Could you let us know if the detailed author contribution should be in the paper, or if it all good as it is now ?
