### 2025-11-05 colmduff/OptiGob#19 @colmduff
I have updated the package to use the validation combinations for input validation as suggested here.  I have also added a logger.
