### 2025-11-05 abbe37f
bug fixes, added logging, input validation
