+++ tests/test_input_validation.py
@@ -1,0 +1,514 @@
+"""
+test_input_validation.py
+========================
+
+This test module verifies the input parameter validation functionality in OptiGobDataManager.
+
+It tests validation for three parameter groups:
+1. Forest parameters (afforestation_rate, broadleaf_fraction, organic_soil_fraction, forest_harvest_intensity)
+2. Organic soil parameters (wetland_restored_frac, organic_soil_under_grass_frac)
+3. Abatement/productivity parameters (abatement_type, abatement_scenario)
+
+The validation ensures that only parameter combinations present in the database are accepted.
+"""
+
+from pathlib import Path
+import pytest
+from optigob.resource_manager.optigob_data_manager import OptiGobDataManager
+from optigob.input_helper import InputHelper
+from optigob.logger import configure_logging
+import logging
+
+LOGPATH = str(Path('.') / 'data' / 'logs' / 'input_validation.log')
+
+# Log to both console and file
+configure_logging(
+    level=logging.INFO,
+    log_to_file=str(LOGPATH)
+)
+
+class TestForestParameterValidation:
+    """Test validation of forest parameter combinations."""
+
+    def get_base_params(self):
+        """Return minimal valid parameters for testing."""
+        return {
+            'AR': 5,
+            'split_gas': False,
+            'target_year': 2030,
+            'abatement_type': 'frontier',
+            'abatement_scenario': 9,
+            'livestock_ratio_type': 'dairy_per_beef',
+            'livestock_ratio_value': 10,
+            'baseline_year': 2020,
+            'baseline_dairy_pop': 156.8,
+            'baseline_beef_pop': 98.4,
+        }
+
+    def test_valid_forest_combination(self):
+        """Test that a valid forest combination is accepted."""
+        params = self.get_base_params()
+        params.update({
+            'afforestation_rate_kha_per_year': 16.0,
+            'broadleaf_fraction': 0.3,
+            'organic_soil_fraction': 0.0,
+            'forest_harvest_intensity': 'low'
+        })
+        # Should not raise
+        data_manager = OptiGobDataManager(params)
+        assert data_manager is not None
+
+    def test_invalid_forest_combination_wrong_organic_soil_frac(self):
+        """Test that an invalid forest combination (wrong organic_soil_fraction) is rejected."""
+        params = self.get_base_params()
+        params.update({
+            'afforestation_rate_kha_per_year': 16.0,
+            'broadleaf_fraction': 0.3,
+            'organic_soil_fraction': 0.5,  # Invalid - not in database
+            'forest_harvest_intensity': 'low'
+        })
+        with pytest.raises(ValueError) as exc_info:
+            OptiGobDataManager(params)
+        assert "Invalid Forest Parameter Combination" in str(exc_info.value)
+        assert "organic_soil_fraction: 0.5" in str(exc_info.value)
+
+    def test_invalid_forest_combination_wrong_broadleaf(self):
+        """Test that an invalid forest combination (wrong broadleaf_fraction) is rejected."""
+        params = self.get_base_params()
+        params.update({
+            'afforestation_rate_kha_per_year': 8.0,
+            'broadleaf_fraction': 0.7,  # Invalid - not in database
+            'organic_soil_fraction': 0.15,
+            'forest_harvest_intensity': 'high'
+        })
+        with pytest.raises(ValueError) as exc_info:
+            OptiGobDataManager(params)
+        assert "Invalid Forest Parameter Combination" in str(exc_info.value)
+        assert "broadleaf_fraction: 0.7" in str(exc_info.value)
+
+    def test_invalid_forest_combination_wrong_harvest(self):
+        """Test that an invalid forest combination (wrong harvest intensity) is rejected."""
[…]
