### 2025-11-05 9005a81
bug fixes, added logging, input validation
