### 2025-11-05 feffb6b
bug fixes, added logging, input validation
