### 2025-11-05 #9271 @mengqi-z
Thanks for raising this concern, @lionel42. It’s up to the authors to determine appropriate credit in line with JOSS’s authorship policy. Just a note that providing only financial support or general supervision does not qualify someone for co-authorship. Based on @colmduff’s statement, contributions from other co-authors are sufficient for co-authorship given that they were substantially involved in the development. It is not required to include detailed author contribution statements in the paper.
