+++ joss_paper/paper/paper.md
@@ -1,0 +1,80 @@
+---
+title: 'OptiGob: A decision support tool for Agriculture, Forestry and Other Land Use transitions'
+tags:
+  - Python
+  - Land use change
+  - climate change
+  - environmental impact
+authors:
+  - name: Colm Duffy
+    orcid: 0000-0002-0076-6749
+    corresponding: true 
+    affiliation: 1
+  - name: Daniel Henn
+    orcid: 0000-0002-2249-158X
+    affiliation: 1
+  - name: Takfarinas Saber
+    orcid: 0000-0003-2958-7979
+    affiliation: 2
+  - name: David Styles
+    orcid: 0000-0003-4185-4478
+    affiliation: 1
+affiliations:
+  - name: School of Biological and Chemical Sciences and Ryan Institute, University of Galway, H91 REW4, Ireland
+    index: 1
+  - name: Lero, School of Computer Science, University of Galway, Galway, Ireland
+    index: 2
+
+date: 10 July 2025
+bibliography: paper.bib
+
+---
+# Summary
+
+`OptiGob` is a Python-based tool designed to explore configurations of Ireland’s agriculture, forestry, and other land use (AFOLU) sectors. The main purpose is to assist users in the assessment of environmental and economic impact pathways, based on different land use transition pathways under varying assumptions about agricultural, afforestation, emissions abatement, and carbon dioxide removal (CDR) strategies. `OptiGob` combines data outputs from the GOBLIN Lite for agriculture and land use, from FERS-CBM for forestry, and from the LCAD2.0 for anaerobic digestion, to generate scenarios that respect biophysical constraints. `OptiGob` provides a flexible, customisable, tool for researchers, policymakers, and educators to explore environmental and economic trade-offs associated with land use transition pathways.
+
+# Statement of need
+
+Ireland's agricultural landscape is dominated by grassland, supporting extensive dairy and suckler beef production. Investment in Ireland's bio-economy requires substantial changes to the AFOLU sector. `OptiGob` allows users to manipulate critical AFOLU levers that determine transition pathways to explore their impacts. `OptiGob` utilises data from GOBLIN Lite [@Duffy:2024; @DuffyB:2022], FERS-CBM [@Black:2025; @Kurz:2008], and LCAD2.0 [@Martinez-Acre:2025] to calculate outputs from agriculture, forestry, and anaerobic digestion.
+
+`OptiGob` estimates the net greenhouse gas (GHG) emissions from land use change and estimates available emissions and land budget for grass-based livestock (Dairy and Suckler cow) production. `Pyomo` [@bynum2021pyomo; @hart2011pyomo] is used to optimise livestock populations, while respecting area commitment (afforestation, anaerobic digestion, BECCS (Bioenergy with Carbon Capture and Storage), protein crops) and emission (CO~2~e in the case of net-zero, or CH~4~ alongside net-zero CO~2~e for N~2~O and CO~2~ under a split gas target) constraints.
+
+The GOBLIN framework has been applied in recent studies of net-zero pathways for AFOLU [@Henn:2025; @Bishop:2024; @DuffyB:2022]. `OptiGob` builds upon this framework, offering insight into synergies and trade-offs with reduced overhead and complexity.
+
+# Model Overview
+
[…]
