+++ joss_paper/paper/paper.bib
@@ -1,129 +1,0 @@
-
-@article{DuffyB:2022,
-  	doi = {10.5194/gmd-15-2239-2022},
-  	Archiveprefix = {gmd},
-  	Author = {{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D.},
-  	Journal = {Geoscientific Model Development},
-  	Title = {{GOBLIN version 1.0: a land balance model to identify national agriculture and land use pathways to climate neutrality via backcasting}},
-  	Year = 2022
-}
-@article{Duffy:2024,
-  	doi = {10.21105/joss.06732},
-  	Archiveprefix = {joss},
-  	Author = {{Duffy}, C. and {Henn}, D. and {Prudhomme}, R. and {Styles}, D.},
-  	Journal = {The Journal of Open Source Software},
-  	Title = {{GOBLIN Lite: A National Land Balance Model for Assessment of Climate Mitigation Pathways for Ireland.}},
-  	Year = 2024
-}
-@article{Kurz:2008,
-  	doi = {10.1016/j.ecolmodel.2008.10.018},
-  	Archiveprefix = {ecolmodel},
-  	Author = {{Kurz}, W.A. and {Dymond}, C.C. and {White}, T.M. and {Stinson}, G. and {Shaw}, C.H. and {Rampley}, G.J. and {Smith}, C. and {Simpson}, B.N. and {Neilson}, E.T. and {Trofymow}, J.A. and {Metsaranta}, J. and {Apps}, M.J.},
-  	Journal = {Ecological Modelling},
-  	Month = mar,
-  	Title = {{CBM-CFS3: A model of carbon-dynamics in forestry and land-use change implementing {IPCC} standards}},
-  	Year = 2008
-}
-@article{Black:2025,
-  	doi = {10.1186/s13021-025-00302-7},
-  	Author = {{Black}, K. and {McCullagh}, A. and {Redmond}, J.},
-  	Journal = {Carbon Balance Management},
-  	Title = {{Evaluation of climate change mitigation strategies for Irish forests using the CBM-CFS3 model}},
-  	Year = 2025
-}
-@article{Martinez-Acre:2025,
-  	doi = {},
-  	Author = {{Martinez-Acre}, A.},
-  	Journal = {Under review},
-  	Title = {{Critical evaluation of prospective biorefinery configurations to deliver a circular, climate neutral economy}},
-  	Year = 2025
-}
-@article{Bishop:2024,
-  	doi = {10.1038/s43247-024-01275-0},
-  	Author = {{Bishop}, G. and {Duffy}, C. and {Prudhomme}, R. and {Cowie}, A. and {O'Donoghue}, C. and {Cain}, M. and {Lanigan}, G.J. and {Styles}, D.},
-  	Journal = {Communications Earth \& Environment},
-  	Title = {{Defining national {Net-Zero} goals is critical for food and land use policy}},
-  	Year = 2024
-}
-@article{DuffyA:2022,
-  	doi = {10.1038/s41893-022-00946-0},
-  	Author = {{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {Iannetta}, P.P.M. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D.},
-  	Journal = {Nature Sustainability},
-  	Title = {{Randomized national land management strategies for {Net-Zero} emissions}},
-  	Year = 2022
-}
-@misc{NIR,
-  author = {Duffy, P. and Black, K. and Fahey, D. and Hyde, B. and Kehoe, A. and Monaghan, S. and Murphy, J. and Ryan, A.M. and Ponzi, J.},
[…]
