### 2025-11-05 colmduff/OptiGob#20 @colmduff
I will try to summarise the changes that I have address here

- All textual changes request by the reviewer have been made.
- I have addressed comments related to the Figure 1, including the text size, and adding additional text to clearly define what the arrows represent.
- I have added a line about the database, but this is generated outside of the package, and is just used by the package. So I have not added it to Figure 1.
- I have not added bullet points as a personal preference, I don't use them and don't particulary like seeing them in journal articles, though I do appreciate where the reviewer was coming from here, and I agree that they are good in a documentation setting.
- I have added a jupyter notebook file (instead of marimo, which is just a fancy new notebook that you don't have to reload everytime) for the reviewer, which should allow for the execution and recreation of the output.
- I have removed the functionality section and encorporated relevent pieces in the methods.
- I have created a Limitations.md and added it to the documentation.

I will close this now, as I think this addresses any major points, but I am happy to address anyting else the reviewer feels is outstanding.

Thank you again for a comprehensive and helpful review. It has certianly made the package better, and challenged me to think about areas of the code I had not paid a huge amount of attention to.
