### 2025-11-06 #9271 @mengqi-z
@colmduff - Thanks for the update! Just to clarify, you don’t need to release every version to Zenodo for JOSS purposes. We only require a Zenodo (or other DOI-issuing archive) record once the paper is accepted and includes the latest changes. Of course, you’re welcome to create a Zenodo archive for your software releases if you’d like.
