### 2026-02-13 colmduff/OptiGob#25 @quaz115
Installation failure of OptiGob due to cplex dependency (refinement needed in Installation guide)
[@openjournals/joss-reviews#9271](https://github.com/openjournals/joss-reviews/issues/9271)
I was unable to install the package using `pip install optigob`. Both available versions (0.1.3 and 0.1.4) fail with a `ResolutionImpossible` error caused by the hard dependency on `cplex>=22.1.2.0,<23.0.0.0`:

Please refine and address this in the Installation guide:  https://optigob.readthedocs.io/en/latest/installation.html#requirements
```
pip install optigob
Collecting optigob
  Downloading optigob-0.1.4-py3-none-any.whl.metadata (11 kB)
INFO: pip is looking at multiple versions of optigob to determine which version is compatible with other requirements. This could take a while.
  Downloading optigob-0.1.3-py3-none-any.whl.metadata (10 kB)
ERROR: Cannot install optigob==0.1.3 and optigob==0.1.4 because these package versions have conflicting dependencies.

The conflict is caused by:
    optigob 0.1.4 depends on cplex<23.0.0.0 and >=22.1.2.0
    optigob 0.1.3 depends on cplex<23.0.0.0 and >=22.1.2.0

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip to attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/topics/dependency-resolution/#dealing-with-dependency-conflicts

```

The `cplex` package (IBM's CPLEX optimizer) is a commercial/proprietary solver with restricted availability on PyPI — it is not installable on most platforms and Python versions without a valid IBM license or specific environment setup. This makes the package effectively **uninstallable for the general user** via a simple `pip install`, which is a JOSS requirement.

**Recommendations:**

1. **Make `cplex` an optional dependency** rather than a hard requirement. Consider moving it to an extras group (e.g., `pip install optigob[cplex]`) so the base package can be installed without it.
2. **Support an open-source solver alternative** such as [GLPK](https://www.gnu.org/software/glpk/), [HiGHS](https://highs.dev/), or [CBC](https://github.com/coin-or/Cbc) via Pyomo's solver interface, so users without a CPLEX license can still use the tool.
3. **Document the solver requirement clearly** in the installation instructions, including how to obtain CPLEX (e.g., via the IBM Academic Initiative for researchers) and any platform/Python version constraints.
4. At minimum, ensure `pip install optigob` succeeds and the package can be imported, even if full solver functionality requires additional setup.
