### 2026-02-13 colmduff/OptiGob#25 @quaz115
cplex issue also persists with poetry update
```
poetry update
The "poetry.dev-dependencies" section is deprecated and will be removed in a future version. Use "poetry.group.dev.dependencies" instead.
Updating dependencies
Resolving dependencies... (7.7s)

Package operations: 1 install, 34 updates, 1 removal

  - Removing sniffio (1.3.1)
  - Updating rpds-py (0.28.0 -> 0.30.0)
  - Updating pycparser (2.23 -> 3.0)
  - Updating platformdirs (4.5.0 -> 4.7.0)
  - Updating tzdata (2025.2 -> 2025.3)
  - Updating jsonschema (4.25.1 -> 4.26.0)
  - Updating tornado (6.5.2 -> 6.5.4)
  - Updating jupyter-client (8.6.3 -> 8.8.0)
  - Updating soupsieve (2.8 -> 2.8.3)
  - Updating asttokens (3.0.0 -> 3.0.1)
  - Updating beautifulsoup4 (4.14.2 -> 4.14.3)
  - Updating packaging (25.0 -> 26.0)
  - Updating nbclient (0.10.2 -> 0.10.4)
  - Updating mistune (3.1.4 -> 3.2.0)
  - Updating parso (0.8.5 -> 0.8.6)
  - Updating wcwidth (0.2.14 -> 0.6.0)
  - Updating anyio (4.11.0 -> 4.12.1)
  - Updating certifi (2025.10.5 -> 2026.1.4)
  - Updating jupyter-server-terminals (0.5.3 -> 0.5.4)
  - Updating prometheus-client (0.23.1 -> 0.24.1)
  - Updating nbconvert (7.16.6 -> 7.17.0)
  - Updating urllib3 (2.5.0 -> 2.6.3)
  - Updating send2trash (1.8.3 -> 2.1.0)
  - Updating psutil (7.1.3 -> 7.2.2)
  - Updating debugpy (1.8.17 -> 1.8.20)
  - Updating json5 (0.12.1 -> 0.13.0)
  - Updating ipython (9.7.0 -> 9.10.0)
  - Updating babel (2.17.0 -> 2.18.0)
  - Updating greenlet (3.2.4 -> 3.3.1)
  - Updating async-lru (2.0.5 -> 2.1.0)
  - Updating ipykernel (7.1.0 -> 7.2.0)
  - Updating numpy (2.3.4 -> 2.4.2)
  - Updating setuptools (80.9.0 -> 82.0.0)
  - Installing cplex (22.1.2.0): Failed

    | Unable to find installation candidates for cplex (22.1.2.0)
    |
    | This is likely not a Poetry issue.
    |
    |   - 24 candidate(s) were identified for the package
    |   - 24 wheel(s) were skipped as your project's environment does not support the identified abi tags
    |
    | Solutions:
    | Make sure the lockfile is up-to-date. You can try one of the following;
    |
    |     1. Regenerate lockfile: poetry lock --no-cache --regenerate
    |     2. Update package     : poetry update --no-cache cplex
    |
    | If any of those solutions worked, you will have to clear your caches using (poetry cache clear --all).
    |
    | If neither works, please first check to verify that the cplex has published wheels available from your configured source that are compatible with your environment- ie. operating system, architecture (x86_64, arm64 etc.), python interpreter.
    |
    | You can also run your poetry command with -v to see more information.

  - Updating jupyterlab (4.4.10 -> 4.5.4)
  - Updating sqlalchemy (2.0.44 -> 2.0.46)

```
