### 2026-02-13 colmduff/OptiGob#25 @mengqi-z
This is for JOSS paper review openjournals/joss-reviews#9271
(FYI, mentioning `openjournals/joss-reviews#9271` will automatically link the issue to the JOSS review thread, which helps me to track the review progress.)
