### 2026-02-13 colmduff/OptiGob#26 @quaz115
Installation guide missing key steps; all 23 tests fail with "ModuleNotFoundError" after following instructions
https://github.com/openjournals/joss-reviews/issues/9271

The developer installation guide is incomplete — key Poetry steps are missing — and after following the available instructions, all 23 test files fail to collect with `ModuleNotFoundError: No module named 'optigob'`. Additionally, `pyproject.toml` contains a deprecated configuration section.

## Issues

### 1. Installation guide missing critical Poetry steps

The current [INSTALLATION.md](https://github.com/colmduff/OptiGob/blob/main/INSTALLATION.md) does not include several steps needed for a working developer setup. At minimum, the following should be documented explicitly:

- **`poetry install`** — to install the package and its dependencies into the Poetry-managed virtual environment.
- **`eval $(poetry env activate)`** — to activate the Poetry virtual environment in the current shell. The `eval` wrapper is required and easy to miss.
- Any notes about Python version compatibility (the project requires `^3.12` but Poetry may pick up a different interpreter).

### 2. Deprecated `[tool.poetry.dev-dependencies]` section in `pyproject.toml`

Running any Poetry command produces the following warning:

```
The "poetry.dev-dependencies" section is deprecated and will be removed
in a future version. Use "poetry.group.dev.dependencies" instead.
```

The empty `[tool.poetry.dev-dependencies]` section in `pyproject.toml` should be removed since the project already has the correct `[tool.poetry.group.dev.dependencies]` section defined.

### 3. All 23 tests fail with `ModuleNotFoundError`

After cloning, running `poetry install`, activating the environment, and executing `poetry run pytest`, all 23 test modules fail during collection with:

```
ModuleNotFoundError: No module named 'optigob'
```

This suggests the package is not being installed in development mode, or the `packages` configuration in `pyproject.toml` does not correctly point to the `src/` layout. Since the source code lives under `src/optigob/`, Poetry needs to be configured to find it there (e.g., via `packages = [{include = "optigob", from = "src"}]`).

<summary>Full pytest output (23 errors)</summary>

```
poetry run pytest
collected 0 items / 23 errors

ERROR tests/test_baseline_budget_model.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_baseline_livestock.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_baseline_other_land.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_bioenergy.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_budget_model.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_bugs_enhanced.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_co2e_budget_confirmation.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_database_manager.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_econ_output.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_forest.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_input_helper.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_input_query.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_input_validation.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_land_area_budget_model.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_livestock_budget_model.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_livestock_scaler.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_optigob.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_optimiser.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_otherland.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_protein_crops_budget.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_static_ag.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_substitution.py - ModuleNotFoundError: No module named 'optigob'
ERROR tests/test_zero_budget_message.py - ModuleNotFoundError: No module named 'optigob'

23 errors in 0.56s
```

## Recommended fixes

1. **Remove** the empty `[tool.poetry.dev-dependencies]` section from `pyproject.toml` to eliminate the deprecation warning.
2. **Add** a `packages` directive to `pyproject.toml` if missing:
   ```toml
   packages = [{include = "optigob", from = "src"}]
   ```
3. **Update INSTALLATION.md** with explicit developer setup steps, e.g.:
   ```bash
   git clone https://github.com/colmduff/OptiGob.git
   cd OptiGob
   poetry install
   eval $(poetry env activate)
   poetry run pytest
   ```
4. Verify that `poetry run pytest` passes all 23 tests before releasing.
