### 2026-02-13 colmduff/OptiGob#26 @colmduff
Thank you for this detailed feedback. It appears the ModuleNotFoundError during testing occurred because the package was not installed in editable mode within the environment.

I can make some changes that may assist.

- Add packages include in pyproject.toml
- change depreciated element to [tool.poetry.group.dev.dependencies]
- I will update the instructions
