### 2026-02-13 colmduff/OptiGob#25 @colmduff
Thank you for the feedback. I am  examining the dependency structure for broader compatibility. However, I would like to clarify a few points regarding the cplex requirement:

- The cplex package used here is the Community Edition, which is free, and publicly available on [PyPI](https://pypi.org/project/cplex/).

While the error you encountered suggests a platform mismatch in this case, I would point the reviewer to the 24 available wheels on the [Community Edition on PyPI](https://pypi.org/project/cplex/#files). This package is readily available to the vast majoirty of users.

Proposed Actions:

- I can make cplex solver optional
- I will verify compatibility with alternative solvers

These changes should remove any issues.
