### 2026-02-16 colmduff/OptiGob#25 @colmduff
I have updated the package making the solver optional

- Optional solver == "highs"
- User can use this or another solver, must specifiy solver name in the SIT.yaml
- Updated Installation.md
- Updated Project.tml
- Updated README.md
- Updated livestock_optimisation.py
- optigob_data_manager.py

Still works with cplex_direct, have given the user the greater degree of flexibility, and there is not block on package installation.
