### 2026-02-17 colmduff/OptiGob#26 @quaz115
https://github.com/openjournals/joss-reviews/issues/9271  @colmduff   I will note that i was successful with conda installation but python=3.12 not 3.10  and for poetry install its better to follow these steps:   `**poetry cache clear . --all** (if only "poetry install" stalls - can happen for login node of an HPC cluster)
Delete 142 entries? (yes/no) [yes] yes
(optigob) |19:14:01|Quazi.Rasool@hfe09:[OptiGob]>
(optigob) |19:14:05|Quazi.Rasool@hfe09:[OptiGob]> **poetry config installer.max-workers 1**
The "poetry.dev-dependencies" section is deprecated and will be removed in a future version. Use "poetry.group.dev.dependencies" instead.

(optigob) |19:14:42|Quazi.Rasool@hfe09:[OptiGob]> **poetry install**
The "poetry.dev-dependencies" section is deprecated and will be removed in a future version. Use "poetry.group.dev.dependencies" instead.
Installing dependencies from lock file

Package operations: 21 installs, 2 updates, 0 removals

  - Installing referencing (0.37.0)
  - Installing rfc3339-validator (0.1.4)
  - Installing rfc3986-validator (0.1.1)
  - Installing rfc3987-syntax (1.1.0)
  - Installing rpds-py (0.28.0)
  - Installing send2trash (1.8.3)
  - Downgrading setuptools (82.0.0 -> 80.9.0): Installing...
Installing /scratch4/BMC/rcm1/qrasool/anaconda3/envs/optigob/lib/python3.12/site-packages/distutils-precedence.pth over existing   - Downgrading setuptools (82.0.0 -> 80.9.0)
  - Installing six (1.17.0)
  - Installing sniffio (1.3.1)
  - Installing soupsieve (2.8)
  - Installing sqlalchemy (2.0.44)
  - Installing stack-data (0.6.3)
  - Installing terminado (0.18.1)
  - Installing tinycss2 (1.4.0)
  - Installing tornado (6.5.2)
  - Installing traitlets (5.14.3)
  - Installing tzdata (2025.2)
  - Installing uri-template (1.3.0)
  - Downgrading urllib3 (2.6.3 -> 2.5.0)
  - Installing wcwidth (0.2.14)
  - Installing webcolors (25.10.0)
  - Installing webencodings (0.5.1)
  - Installing websocket-client (1.9.0)

Installing the current project: optigob (0.1.4)
(optigob) |19:14:53|Quazi.Rasool@hfe09:[OptiGob]> **poetry run python -c "import pyomo; print(pyomo.__version__)"**
The "poetry.dev-dependencies" section is deprecated and will be removed in a future version. Use "poetry.group.dev.dependencies" instead.
6.9.5
`

**And eventually was able to run the example.py:**

```
(optigob) |19:32:54|Quazi.Rasool@hfe09:[OptiGob]> cd tests/
(optigob) |19:39:24|Quazi.Rasool@hfe09:[tests]> poetry run python example.py
The "poetry.dev-dependencies" section is deprecated and will be removed in a future version. Use "poetry.group.dev.dependencies" instead.
##################################################
OptiGob Budget Model Input Combinations

Input type: forest
  affor_rate_kha-yr | broadleaf_frac | organic_soil_frac | forest_harvest_intensity
  -----------------------------------------------------------------------------------
  2 | 0.5 | 0.15 | high
  8 | 0.5 | 0.15 | high
  8 | 0.5 | 0.15 | low
  8 | 0.3 | 0.15 | high
  8 | 0.3 | 0.15 | low
  16 | 0.5 | 0.15 | low
  16 | 0.3 | 0.0 | high
  16 | 0.3 | 0.0 | low
  25 | 0.5 | 0.15 | high
  25 | 0.5 | 0.15 | low
  25 | 0.3 | 0.0 | high
  25 | 0.3 | 0.0 | low

Input type: organic_soil
  wetland_restored_frac | organic_soil_under_grass_frac
  -------------------------------------------------------
  0.5 | 0.0
  0.9 | 0.5
  0.9 | 0.9

Input type: abatement_and_productivity
  abatement | scenario
  ----------------------
  baseline | 1
  baseline | 2
  baseline | 3
  macc | 4
  macc | 5
  macc | 6
  frontier | 7
  frontier | 8
  frontier | 9
##################################################
OptiGob Budget Model Input Combinations
##################################################
GHG Emissions by Sector
{'baseline': {'agriculture': 21147.772800304, 'existing_forest': -2082.0, 'afforestation': 0, 'hwp': 0, 'other_land_use': 5947, 'ad': 0, 'beccs': 0}, 'scenario': {'agriculture': 2186.5649762, 'existing_forest': 1692.0, 'afforestation': -278.868963, 'hwp': -1956, 'other_land_use': 5075, 'ad': 141.1, 'beccs': 0.0}}
                   baseline     scenario
agriculture      21147.7728  2186.564976
existing_forest  -2082.0000  1692.000000
afforestation        0.0000  -278.868963
hwp                  0.0000 -1956.000000
other_land_use    5947.0000  5075.000000
ad                   0.0000   141.100000
beccs                0.0000     0.000000
##################################################
Aggregated Total Land Area by Sector
{'baseline': {'agriculture': 4225153.146096, 'afforested': 0, 'existing_forest': 773254, 'other_land_use': 1433000, 'ad': 0, 'protein_crops': 0, 'beccs_willow': 0}, 'scenario': {'agriculture': 1100562.56, 'afforested': 48640, 'existing_forest': 781254, 'other_land_use': 1433000, 'ad': 130000.0, 'protein_crops': 0, 'beccs_willow': 150000.0}}
                     baseline    scenario
agriculture      4.225153e+06  1100562.56
afforested       0.000000e+00    48640.00
existing_forest  7.732540e+05   781254.00
other_land_use   1.433000e+06  1433000.00
ad               0.000000e+00   130000.00
protein_crops    0.000000e+00        0.00
beccs_willow     0.000000e+00   150000.00
##################################################
Protein by Sector
{'baseline': {'pig_and_poultry': 132000000.0, 'sheep': 14332176.614848224, 'beef': 155417070.62386402, 'milk': 297463320.00000006, 'protein_crops': 0}, 'scenario': {'pig_and_poultry': 158400000.0, 'sheep': 14332176.614848224, 'beef': 0.0, 'milk': 0.0, 'protein_crops': 0}}
                     baseline      scenario
pig_and_poultry  1.320000e+08  1.584000e+08
sheep            1.433218e+07  1.433218e+07
beef             1.554171e+08  0.000000e+00
milk             2.974633e+08  0.000000e+00
protein_crops    0.000000e+00  0.000000e+00
##################################################
Area by Sector
{'baseline': {'dairy': 980841.4292544001, 'beef': 2066291.4368416, 'sheep': 828020.28, 'pig_and_poultry': 0.0, 'static_crops': 350000, 'protein_crops': 0, 'beccs_willow': 0, 'anaerobic_digestion': 0, 'managed forest': 773254, 'afforestation': 0, 'other_land_use': 1433000, 'available_area': 0}, 'scenario': {'dairy': 0.0, 'beef': 0.0, 'sheep': 750562.56, 'pig_and_poultry': 0.0, 'static_crops': 350000.0, 'protein_crops': 0, 'beccs_willow': 150000.0, 'anaerobic_digestion': 130000.0, 'managed forest': 781254, 'afforestation': 48640, 'other_land_use': 1433000, 'available_area': 2787950.5860960004}}
                         baseline      scenario
dairy                9.808414e+05  0.000000e+00
beef                 2.066291e+06  0.000000e+00
sheep                8.280203e+05  7.505626e+05
pig_and_poultry      0.000000e+00  0.000000e+00
static_crops         3.500000e+05  3.500000e+05
protein_crops        0.000000e+00  0.000000e+00
beccs_willow         0.000000e+00  1.500000e+05
anaerobic_digestion  0.000000e+00  1.300000e+05
managed forest       7.732540e+05  7.812540e+05
afforestation        0.000000e+00  4.864000e+04
other_land_use       1.433000e+06  1.433000e+06
available_area       0.000000e+00  2.787951e+06
##################################################
High Nature Value (HNV) Land Area by Sector
{'baseline': {'agriculture': 681876.1741560001, 'afforested': 0, 'existing_forest': 255174, 'wetland': 884000, 'rewetted_wetland': 85000, 'organic_soil': 85000, 'beccs': 0}, 'scenario': {'agriculture': 0.0, 'afforested': 14592.0, 'existing_forest': 257814, 'wetland': 884000, 'rewetted_wetland': 198000, 'organic_soil': 198000, 'beccs': 150000.0}}
                       baseline  scenario
agriculture       681876.174156       0.0
afforested             0.000000   14592.0
existing_forest   255174.000000  257814.0
wetland           884000.000000  884000.0
rewetted_wetland   85000.000000  198000.0
organic_soil       85000.000000  198000.0
beccs                  0.000000  150000.0
##################################################
Bioenergy by Sector
{'baseline': {'ad': 0, 'willow_biomass': 0, 'forest_biomass': 0}, 'scenario': {'ad': 5700000, 'willow_biomass': 6605280.0, 'forest_biomass': 7316474.089}}
                baseline     scenario
ad                     0  5700000.000
willow_biomass         0  6605280.000
forest_biomass         0  7316474.089
##################################################
HWP
{'baseline': {'hwp': 0}, 'scenario': {'hwp': 6586461}}
     baseline  scenario
hwp         0   6586461
##################################################
Substitution
{'ad_substitution': -1495.525039, 'forest_substitution': -5496, 'willow_substitution': -954.46296}
                        scenario
ad_substitution     -1495.525039
forest_substitution -5496.000000
willow_substitution  -954.462960
##################################################
NZ Status
False
total emissions co2e: 6859.796013200001 kt
##################################################
Livestock Population
{'baseline': {'dairy': 1568000.0, 'beef': 984000.0}, 'scenario': {'dairy': 0.0, 'beef': 0.0}}
        baseline  scenario
dairy  1568000.0       0.0
beef    984000.0       0.0
##################################################
Livestock CH4 Emissions budget
283.878965804
##################################################
Livestock CO2e Emissions budget
0
##################################################
AREA comparison
                         baseline      scenario
dairy                9.808414e+05  0.000000e+00
beef                 2.066291e+06  0.000000e+00
sheep                8.280203e+05  7.505626e+05
pig_and_poultry      0.000000e+00  0.000000e+00
static_crops         3.500000e+05  3.500000e+05
protein_crops        0.000000e+00  0.000000e+00
beccs_willow         0.000000e+00  1.500000e+05
anaerobic_digestion  0.000000e+00  1.300000e+05
managed forest       7.732540e+05  7.812540e+05
afforestation        0.000000e+00  4.864000e+04
other_land_use       1.433000e+06  1.433000e+06
available_area       0.000000e+00  2.787951e+06

Sum of each column:
baseline    6.431407e+06
scenario    6.431407e+06
dtype: float64

```
