### 2026-02-17 colmduff/OptiGob#27 @quaz115
Automated Tests —  Cannot check off yet (3 automated tests failing)
https://github.com/openjournals/joss-reviews/issues/9271
---

## Automated tests: Failures due to path inconsistencies and unimplemented assertions

### Summary

**3 test scripts fail when run locally.** The failures fall into two categories:
1. **Inconsistent working directory assumptions** causing `FileNotFoundError`
2. **Tests asserting behavior that is not implemented** in the current codebase

---

### 1. `FileNotFoundError` in `test_zero_budget_message.py` and `test_co2e_budget_confirmation.py`

Both files reference test data using paths relative to the **repository root**:

```python
data_manager = OptiGobDataManager('./tests/data/sip_bug1.yaml')
```

However, the CI workflow (`.github/workflows/test.yml`) and the `tests/Makefile` run tests from **within the `tests/` directory**:

```yaml
# .github/workflows/test.yml
- name: Run tests via Makefile in tests folder
  working-directory: tests
  run: poetry run make test
```

This means `./tests/data/sip_bug1.yaml` resolves to `tests/tests/data/sip_bug1.yaml`, which doesn't exist.

**Other test files** (e.g., `test_forest.py`, `test_optigob.py`, `test_static_ag.py`) correctly use paths relative to the `tests/` directory:

```python
self.data = './data/sip.json'
```

**Reproduction:**

```
(optigob) tests> poetry run python test_zero_budget_message.py

FileNotFoundError: [Errno 2] No such file or directory: './tests/data/sip_bug1.yaml'
```

**Suggested fix:** Change the paths in `test_zero_budget_message.py` and `test_co2e_budget_confirmation.py` from:

```python
OptiGobDataManager('./tests/data/sip_bug1.yaml')
```

to:

```python
OptiGobDataManager('./data/sip_bug1.yaml')
```

---

### 2. `test_bugs_enhanced.py` — Assertions for unimplemented behavior

Running `test_bugs_enhanced.py` produces the following failures:

```
Testing Bug 1 (FIXED): Input validation + zero budget info
✗ No warning raised (expected one)
  Zero budget info message checks:
    ✗ INFO header
    ✗ Net emissions explanation
    ✗ Budget breakdown
    ✗ Zero livestock result
✗ Bug 1 PARTIALLY FIXED: Info message checks failed

Testing Bug 2 (FIXED): Pre-flight CH4 check
✗ Bug 2 test FAILED: Expected ValueError but none was raised
```

- **Bug 1:** The test captures `sys.stdout` and checks for printed info messages (e.g., `"INFO: Zero CO2e budget for livestock"`, `"NET EMISSIONS"`). The captured output is empty, suggesting the software uses `logging` rather than `print()`, so stdout capture doesn't work. Alternatively, the info message logic may not be implemented.

- **Bug 2:** The test expects a `ValueError` to be raised as a pre-flight CH4 budget check, but no exception is raised, indicating this validation is not yet implemented.

---

### 3. CI may not catch these failures

The CI workflow uses `unittest discover`:

```makefile
# tests/Makefile
test:
	@python -m unittest discover -s . -p "*.py"
```

- `test_bugs_enhanced.py` and `test_zero_budget_message.py` define **standalone functions** (not `unittest.TestCase` subclasses), so `unittest discover` will **not** pick them up and they will pass silently.
- `test_co2e_budget_confirmation.py` runs at module level (`if __name__ == "__main__"` with top-level calls), which also won't be discovered by `unittest discover`.

This means the CI may report all tests as passing even though these scripts fail when run directly.

---

### Suggested actions

1. Fix the relative paths in `test_zero_budget_message.py` and `test_co2e_budget_confirmation.py` to use `'./data/...'` instead of `'./tests/data/...'`.
2. Either implement the expected behavior (info messages, CH4 pre-flight validation) or update the tests to match the current behavior.
3. Convert standalone test functions to `unittest.TestCase` classes so they are picked up by `unittest discover` in CI, or switch to a test runner like `pytest` that discovers standalone `test_` functions.
