### 2026-02-17 #9271 @quaz115
https://github.com/openjournals/joss-reviews/issues/9271    @colmduff     For  ✅ **Refernces** please address these:

1. Martinez-Acre:2025 has an empty doi field (doi = {},) and lists the journal as "Under review"? Is there a preprint or DOI available? JOSS generally expects references to be verifiable.
2. Missing DOI for these as well:
```
Hart, W. E., Watson, J.-P., & Woodruff, D. L. (2011). Pyomo: Modeling and solving mathematical programs in python. Mathematical Programming Computation, 3(3), 219–260.

Bynum, M. L., Hackebeil, G. A., Hart, W. E., Laird, C. D., Nicholson, B. L., Siirola, J. D., Watson, J.-P., & Woodruff, D. L. (2021). Pyomo–optimization modeling in python (Third, Vol. 67). Springer Science & Business Media.
```
