### 2026-02-17 #9271 @quaz115
https://github.com/openjournals/joss-reviews/issues/9271 @colmduff
The paper is well-structured and generally clear, but I noticed a few spots that could use `minor fixes`:
However For ✅**Quality of writing**, some minor edits needed:

1. In Summary:  "a flexible, customisable, tool" — _extra comma before "tool" (should be "a flexible, customisable tool")?_
2. In Summary: "based on different land use transition pathways under varying assumptions about **agricultural**....". _Should be_ **"agriculture"**?
3. In the Statement of Need: the phrase "estimates the net greenhouse gas emissions from land use change and estimates available emissions and land budget", repetition of "estimates" a bit awkward — consider rephrasing.
4. The constraint description "(CO₂e in the case of net-zero, or CH₄ alongside net-zero CO₂e for N₂O and CO₂ under a split gas target)" can be rephrased into a more clearer wording?
5. In the Example Output section, "A higher afforestation rate (16 kha per year), with a composition of 70:30 conifer to broadleaf split." reads more as a clause rather than sentence (lacks verb). A few of the surrounding sentences (authors can parse through the pdf and correct such instances) have the same issue — they read as bullet points/clause rather than prose. (mostly one more pass of a copyediting edit would make the paper doc pdf more robust)
