### 2026-02-17 #9271 @quaz115
https://github.com/openjournals/joss-reviews/issues/9271  @colmduff Data Sharing is ✅ (since No original data in the paper), but just confirming — are the underlying datasets (baseline livestock populations, land areas, emissions factors, etc.) bundled in the repo as preconfigured SQLite database (src/optigob/database/) or otherwise accessible at other sources? If so, this item is satisfied.  **However, Authors can feel free (encouraged) to clarify more on Input Data in the [documentation](https://optigob.readthedocs.io/en/latest/index.html)** (like how the paper mentions it as: _data outputs from the GOBLIN Lite for agriculture and land use, from FERS-CBM for forestry, and from the LCAD2.0 for anaerobic digestion,_...)
