### 2026-02-17 colmduff/OptiGob#26 @colmduff
- Added explicit packages = [{include = "optigob", from = "src"}] directive to ensure consistent package discovery across Poetry versions and installation methods
- Confirmed removal of deprecated [tool.poetry.dev-dependencies] section (already using modern [tool.poetry.group.dev.dependencies])
- Updated solver configuration to use [tool.poetry.extras] for optional HiGHS solver dependency

Clarified Poetry environment usage with two explicit options:
- poetry run <command> for single commands (recommended)
- poetry shell for interactive sessions (with exit instructions)
- Confirmed Python 3.12 requirement is correctly documented

Made CPLEX optional.
Added HiGHS as recommended open-source alternative via pip install "optigob[solvers]"

The ModuleNotFoundError issue should be resolved by the explicit packages directive.
