### 2026-02-17 #9271 @colmduff
Yes, the underlying datasets are bundled directly in the repository as a preconfigured SQLite database
  shipped with the package (src/optigob/database/optigob_default_0.1.1.db). This database is installed
  automatically with the package and used by default no additional data download is required.

  To clarify the nature of this data: optigob acts as a simplified surrogate optimisation model that stitches together
   pre-computed outputs from several upstream, sector-specific models:

  - GOBLIN Lite — agriculture and land use emissions/areas
  - FERS-CBM — forestry carbon balance modelling
  - LCAD2.0 — anaerobic digestion

  These upstream models were run across a defined set of valid parameter combinations across three levels of abatement.

  The default database therefore contains derived/modelled data specific to Ireland, not raw observational
  datasets. Users can substitute their own database to change/update the underlying parameters.

Adding a short Data Input section to the docs.
