### 2026-02-17 colmduff/OptiGob#27 @colmduff
I am updating at the moment, but I have not pushed the final edits yet. I was going to push them all together.

I will close this comment one I have updated the joss branch.

FYI, anything that is marked with "bug", will not be pushed in the update. I had used specific inputs to recreate an error and was attempting to trace it. I have put them in the gitignore.
