### 2026-02-17 #9271 @colmduff
> [#9271](https://github.com/openjournals/joss-reviews/issues/9271) [@colmduff](https://github.com/colmduff) The paper's **"Statement of need"** section covers most of JOSS requirements — the research problem is clear (exploring Ireland's AFOLU transition pathways, optimizing livestock under various emissions/land constraints scenarios), the target audience is named ("researchers, policymakers, and educators"), the geographic scope is explicit (Ireland throughout), and there are solid references to prior work (GOBLIN Lite, FERS-CBM, LCAD2.0, Henn et al. 2025, Bishop et al. 2024, etc.).
>
> The one thing I'd like to see improved is the differentiation from existing tools. Right now the paper says OptiGob builds on the GOBLIN framework offering "insight into synergies and trade-offs with reduced overhead and complexity" — but it doesn't really unpack what that means in practice? As a reader, I'm left wondering: what specifically is simpler or faster about using OptiGob versus running GOBLIN Lite directly? And are there other AFOLU scenario-exploration tools out there (outside the GOBLIN family) that someone might compare this to?
>
> Even just few sentence(s) exploring this distinction more concrete would help with ✅ **"Statement of need"**. As an example (but not limited to): "unlike GOBLIN Lite, which requires X, OptiGob allows users to Y" Also for ✅ the **"Statement of need"** **for [Documentation](https://optigob.readthedocs.io/en/latest/index.html)** address this: [documentation landing page](https://optigob.readthedocs.io/en/latest/index.html) jumps straight into module architecture and API details, but doesn't include any high-level context about what problem OptiGob solves or who it's for. While addressing this for the paper author(s) should also this, since someone landing on the docs without having read the paper wouldn't easily understand the purpose of the tool. Even a short introductory paragraph — something like the Summary from the JOSS paper — would help orient new users.
>
> Also helps to ✅ **"State of the field:"** Do the authors describe how this software compares to other commonly-used packages?

---

The Statement of need now clarifies that OptiGob leverages pre-computed outputs from the upstream models (GOBLIN Lite, FERS-CBM, LCAD2.0) without requiring users to run those modelling chains, and notes that to the authors' knowledge this is the first such tool in the Irish context.

The README (which serves as the docs landing page) now opens with a high-level paragraph explaining what OptiGob does, what upstream models it builds on, and who it is for (mirroring the paper's Summary for readers who arrive at the docs directly).
