### 2026-02-17 #9271 @colmduff
> [#9271](https://github.com/openjournals/joss-reviews/issues/9271) [@colmduff](https://github.com/colmduff) Data Sharing is ✅ (since No original data in the paper), but just confirming — are the underlying datasets (baseline livestock populations, land areas, emissions factors, etc.) bundled in the repo as preconfigured SQLite database (src/optigob/database/) or otherwise accessible at other sources? If so, this item is satisfied. **However, Authors can feel free (encouraged) to clarify more on Input Data in the [documentation](https://optigob.readthedocs.io/en/latest/index.html)** (like how the paper mentions it as: _data outputs from the GOBLIN Lite for agriculture and land use, from FERS-CBM for forestry, and from the LCAD2.0 for anaerobic digestion,_...)

---

Yes, the underlying datasets are bundled directly in the repository as a preconfigured SQLite database
shipped with the package (src/optigob/database/optigob_default_0.1.1.db). This database is installed
automatically with the package and used by default no additional data download is required.

To clarify the nature of this data: optigob acts as a simplified surrogate optimisation model that stitches together
pre-computed outputs from several upstream, sector-specific models:

GOBLIN Lite — agriculture and land use emissions/areas
FERS-CBM — forestry carbon balance modelling
LCAD2.0 — anaerobic digestion
These upstream models were run across a defined set of valid parameter combinations across three levels of abatement.

The default database therefore contains derived/modelled data specific to Ireland, not raw observational
datasets. Users can substitute their own database to change/update the underlying parameters.

Adding a short Data Input section to the docs.
