+++ README.md
@@ -8,13 +8,13 @@
 [GitHub](https://github.com/colmduff/OptiGob) |
 [Documentation](https://optigob.readthedocs.io/en/latest/) |
 
-A land use change and environmental assessment tool based on preconfigured data animal population numbers based on negative emissions allowance.
+`OptiGob` is a Python-based decision support tool for exploring Ireland's agriculture, forestry, and other land use (AFOLU) sector under different climate and land use transition pathways. It combines pre-computed scenario data from [GOBLIN Lite](https://github.com/colmduff/goblin_lite) (agriculture and land use), FERS-CBM (forestry carbon balance), and LCAD2.0 (anaerobic digestion) into a single interface that optimises livestock populations subject to land and emissions constraints — without requiring users to run the upstream modelling chains. It is designed for researchers, policymakers, and educators investigating environmental and economic trade-offs associated with net-zero and split-gas emissions targets.
 
 ---
 
 ## High-Level Architecture & Module Interaction
 
-The `optigob` package is a modular land use and environmental assessment framework. It is designed to calculate, aggregate, and analyze greenhouse gas emissions, land use, protein, bioenergy, harvested wood products, and substitution impacts for both baseline and scenario cases. The system is built around a central API (`Optigob`), which coordinates specialized modules for each sector and data type.
+The `optigob` package is a modular land use and environmental assessment framework. It calculates greenhouse gas emissions, land use, protein output, bioenergy, harvested wood products, and substitution impacts for both baseline and scenario cases. The system is built around a central API (`Optigob`), which coordinates specialised modules for each sector and data type.
 
 ### Module Structure
 
@@ -91,9 +91,56 @@ optigob/
 - Substitution logic is centralized in `substitution.py` and called by emissions and output modules as needed.
 - Specialized sector modules (livestock, forest, protein crops, etc.) encapsulate sector-specific calculations and are used by the budget modules.
 
-### Default Database
+### Input Data and Default Database
 
-A default database offers multiple permutations of abatement and productivity scenarios, but certain parameters (e.g., wetland restoration, forestry on organic soils) are limited to predefined combinations. A helper function identifies valid options to prevent unsupported configurations. Custom databases can also be supplied, provided they follow the expected schema.
+`optigob` ships with a pre-built SQLite database (`src/optigob/database/optigob_default_0.1.1.db`) that is installed automatically with the package. **No additional data download is required.**
+
+#### Data provenance
+
+The bundled database does not contain raw observational data. It contains **derived scenario scaler tables** generated by three upstream, sector-specific models:
+
+| Upstream model | Domain | Reference |
+|---|---|---|
+| [GOBLIN Lite](https://github.com/colmduff/goblin_lite) | Agriculture and land use emissions, areas, livestock productivity | Duffy et al. (2022, 2024) |
+| [FERS-CBM](https://www.ucd.ie/forestecosystemresearch/) | Forest carbon balance (afforestation, existing forest, harvested wood) | Black et al. (2025); Kurz et al. (2008) |
+| [LCAD2.0](https://github.com/GOBLIN-Proj/lcad_gob) | Anaerobic digestion area, emissions, and bioenergy output | Martinez-Acre et al. (2025) |
+
+These models were run across a defined set of valid parameter combinations for **Ireland**, and reflect three levels of abatement (Baseline, MACC, Frontier). The resulting scaler tables were stored in the SQLite database. At runtime, `optigob` queries these scalers to parameterise its optimisation model — effectively acting as a surrogate that stitches together pre-computed sectoral outputs without re-running the upstream models.
+
+#### Database contents
+
+The database contains scaler tables for:
+
[…]
