### 2026-02-17 colmduff/OptiGob#27 @colmduff
All github actions tests are passing. Updated test.yml to ensure that optional "highs" solver included.
