### 2026-02-17 #9271 @colmduff
> [#9271](https://github.com/openjournals/joss-reviews/issues/9271) [@colmduff](https://github.com/colmduff) The paper is well-structured and generally clear, but I noticed a few spots that could use `minor fixes`: However For ✅**Quality of writing**, some minor edits needed:
>
> 1. In Summary:  "a flexible, customisable, tool" — _extra comma before "tool" (should be "a flexible, customisable tool")?_
> 2. In Summary: "based on different land use transition pathways under varying assumptions about **agricultural**....". _Should be_ **"agriculture"**?
> 3. In the Statement of Need: the phrase "estimates the net greenhouse gas emissions from land use change and estimates available emissions and land budget", repetition of "estimates" a bit awkward — consider rephrasing.
> 4. The constraint description "(CO₂e in the case of net-zero, or CH₄ alongside net-zero CO₂e for N₂O and CO₂ under a split gas target)" can be rephrased into a more clearer wording?
> 5. In the Example Output section, "A higher afforestation rate (16 kha per year), with a composition of 70:30 conifer to broadleaf split." reads more as a clause rather than sentence (lacks verb). A few of the surrounding sentences (authors can parse through the pdf and correct such instances) have the same issue — they read as bullet points/clause rather than prose. (mostly one more pass of a copyediting edit would make the paper doc pdf more robust)

---
1. changed
2. changed
3. removed second "estimates"
4. I’ve reviewed the suggestion for 'clearer' wording, but I’ve decided to retain the original phrasing for technical precision.
The current description accurately distinguishes between two distinct accounting frameworks: a unified CO₂e net-zero target (where all gases are aggregated) and a split-gas target (which specifically necessitates treating CH₄ separately from the net-zero CO₂e path for N₂O and CO₂). Preservation of this nuance is important for interpretation.
5. A verb ("is applied") has been added to the noted sentence. Regarding the prose style, we have deliberately prioritized information density and economy of language to remain within the strict JOSS word count limits. In the context of a scenario description, we believe a concise, semi-structured format is the most effective way to communicate model parameters to the reader without unnecessary filler.
