### 2026-02-17 472dc78
paper.md and bib.md updates
