+++ joss_paper/paper/paper.bib
@@ -32,10 +32,10 @@
   	Year = 2025
 }
 @article{Martinez-Acre:2025,
-  	doi = {},
-  	Author = {{Martinez-Acre}, A.},
-  	Journal = {Under review},
-  	Title = {{Critical evaluation of prospective biorefinery configurations to deliver a circular, climate neutral economy}},
+  	doi = {10.1016/j.seta.2025.104563},
+  	Author = {{Martinez-Acre}, A. and {Kargupta}, W. and {Girón Domínguez}, C. and {Odey}, E.A. and {McMahon}, H. and {Diaz Huerta}, J. and {Zimmermann}, J. and {Bishop}, G. and {Gaffey}, J. and {Styles}, D.},
+  	Journal = {Sustainable Energy Technologies and Assessments},
+  	Title = {{An integrated technoeconomic and environmental assessment of biomethane production via anaerobic digestion of food waste in Ireland: Updated insights under market volatility}},
   	Year = 2025
 }
 @article{Bishop:2024,
