### 2026-02-17 #9271 @colmduff
Published optigob (0.1.5) to PyPI
