### 2026-02-17 #9271 @quaz115
https://github.com/openjournals/joss-reviews/issues/9271           @colmduff   again pointing to https://github.com/colmduff/OptiGob/issues/27 , one of the 3 failing tests succeeds (had to run from Optiglob home directory, fails when ran after doing `cd tests/` as:  `poetry run python test_co2e_budget_confirmation.py`):

**If they are still BUGS to be resolved please mark them as such and link them in the relevant Issue: https://github.com/colmduff/OptiGob/issues/27 **

 ```
######################################################################
 (optigob)[OptiGob]>**poetry run python tests/test_co2e_budget_confirmation.py**
The "poetry.dev-dependencies" section is deprecated and will be removed in a future version. Use "poetry.group.dev.dependencies" instead.
Parameter Validation Warning:
  split_gas=False but split_gas_frac=0.3
  The split_gas_frac parameter is IGNORED when split_gas=False.
  Set split_gas_frac=0 or remove it to avoid confusion.

Bug 1 Scenario CO2e Budget Analysis

Configuration:
  split_gas: False
  split_gas_frac: 0.3
  target_year: 2030

Emissions Budget Calculation:
  net_zero_budget passed to livestock optimizer: 0.00 kt CO2e
  Raw budget (before abs):                      0.00 kt CO2e

Non-livestock emission components:
  Static agriculture:         2417.95 kt CO2e
  Forest offset:              -542.87 kt CO2e
  Wetland restoration:        5075.00 kt CO2e
  BECCS:                         0.00 kt CO2e
  Protein crops:                 0.00 kt CO2e
  Anaerobic digestion:         141.10 kt CO2e
  ──────────────────────────────────────────────────
  TOTAL non-livestock:        7091.18 kt CO2e

Budget Logic:
  Total is POSITIVE (net emissions) → Budget set to 0

CONFIRMATION: CO2e budget for livestock = 0.00 kt

✓ Confirmed: CO2e budget is exactly 0.00 kt
```

Other 2 tests still failing despite switching to full path of .py and running from `Optigob ` directory instead of `Optigob/tests/`

```
  (optigob)[OptiGob]>**poetry run python tests/test_bugs_enhanced.py**
The "poetry.dev-dependencies" section is deprecated and will be removed in a future version. Use "poetry.group.dev.dependencies" instead.

######################################################################
OPTIGOB BUG FIX VERIFICATION
######################################################################

Testing Bug 1 (FIXED): Input validation + zero budget info
Parameter Validation Warning:
  split_gas=False but split_gas_frac=0.3
  The split_gas_frac parameter is IGNORED when split_gas=False.
  Set split_gas_frac=0 or remove it to avoid confusion.

✗ No warning raised (expected one)

✓ Optimization completed successfully
  Scenario dairy: 0.0
  Scenario beef:  0.0
  → Zero livestock is correct (zero emissions budget)

  Zero budget info message checks:
    ✗ INFO header
    ✗ Net emissions explanation
    ✗ Budget breakdown
    ✗ Zero livestock result

✗ Bug 1 PARTIALLY FIXED: Info message checks failed

  Captured output:

Testing Bug 2 (FIXED): Pre-flight CH4 check

✗ Bug 2 test FAILED: Expected ValueError but none was raised

######################################################################
TEST SUMMARY
######################################################################
  ✗ FAIL: bug1
  ✗ FAIL: bug2

✗ Some tests failed - see details above
```
```

(optigob)[OptiGob]> **poetry run python tests/test_zero_budget_message.py**
The "poetry.dev-dependencies" section is deprecated and will be removed in a future version. Use "poetry.group.dev.dependencies" instead.

Testing Zero CO2e Budget Informational Message
Parameter Validation Warning:
  split_gas=False but split_gas_frac=0.3
  The split_gas_frac parameter is IGNORED when split_gas=False.
  Set split_gas_frac=0 or remove it to avoid confusion.

Captured Output:
──────────────────────────────────────────────────────────────────────

──────────────────────────────────────────────────────────────────────

Verifying message content:
  ✗ INFO header: False
  ✗ Net emissions explanation: False
  ✗ Budget breakdown: False
  ✗ Zero livestock result: False
  ✗ Actionable suggestions: False
  ✗ Wetland restoration: False
  ✗ Static agriculture: False

Population results:
  Dairy population: 0.00
  Beef population: 0.00
Traceback (most recent call last):
  File "/scratch3/BMC/rcm2/qrasool/OptiGob/tests/test_zero_budget_message.py", line 79, in <module>
    test_zero_budget_info_message()
  File "/scratch3/BMC/rcm2/qrasool/OptiGob/tests/test_zero_budget_message.py", line 76, in test_zero_budget_info_message
    raise AssertionError("Some message checks failed")
AssertionError: Some message checks failed
```
