### 2026-02-17 #9271 @colmduff
Regarding the two tests mentioned, these are legacy artifacts that are no longer part of the active test suite.

The Makefile, which governs the official build and the GitHub Actions CI, is the authoritative source for the testing workflow.

That is passing. I have already made the changes requested and bumped the version number, and I have added another wheel to Pypi.

I don't believe an additional version bump or a repository cleanup is necessary for irrelevant files that do not impact the software's functionality or the verified test suite.

I’d prefer to keep the focus on the passing status of the integrated CI.

Those tests were part process to find and address a previous issue.

I would suggest this to be closed and we try and move on and complete this review please.
