### 2026-02-18 #9271 @lionel42
@mengqi-z  Thanks for letting me know the review is almost at the end.

the only point of concern is one that we already raised and discussed

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

The authors claim is

> To the authors’ knowledge, OptiGob represents the first attempt to do so in the Irish context

In this case, it would be good to cite similar software that apply in other contexts.

Also, the authors have claimed in a previous comment:

>  In the next update, we will add: A tutorial showing how to modify or replace the SQLite database to adapt the model to another region or dataset.

I am not aware of such a tutorial on the current release.

My main fear is that since this software was developed for such a specific use case, such that the project will be discontinued and not maintained in the future.

There exist a lot of similar tools, and the authors should make the effort to at least cite some of them in their statement of need and explain why they needed to create a new one.

Once this is addressed, the submission can be accepted.
