### 2026-02-20 #9271 @mengqi-z
@lionel42 - Thank you for summarizing the remaining issues and for raising your concerns.

@colmduff - JOSS requires the paper to describe the current state of the field relevant to this software. As the reviewer suggested, please cite comparable tools and state their similarities and differences. It is important to explain the specific contributions and added value of this software compared to others. Please update the thread once you have addressed these comments. Thanks!
