+++ joss_paper/paper/paper.bib
@@ -129,3 +129,39 @@
   note = {Prepared by Teagasc, Oak Park, Carlow},
   url = {https://teagasc.ie/wp-content/uploads/media/website/environment/climate-action/climate-centre/MACC-2023.pdf}
 }
+
+@article{Leip:2008,
+  doi = {https://doi.org/10.5194/bg-5-73-2008},
+  title = {Linking an economic model for European agriculture with a mechanistic model to estimate nitrogen and carbon losses from arable soils in Europe},
+  author = {Leip, A. and Marchi, G. and Koeble, R. and Kempen, M. and Britz, W. and Li, C.},
+  year = {2008},
+  publisher = {Copernicus GmbH},
+  journal = {Biogeosciences}
+}
+
+@article{Havlík:2010,
+  doi = {https://doi.org/10.1016/j.enpol.2010.03.030},
+  title = {Global land-use implications of first and second generation biofuel targets},
+  author = {Havlík, Petr and Schneider, Uwe A. and Schmid, Erwin and Böttcher, Hannes and Fritz, Steffen and Skalský, Rastislav and Aoki, Kentaro and De Cara, Stéphane and Kindermann, Georg and Kraxner, Florian and Leduc, Sylvain and McCallum, Ian and Mosnier, Aline and Sauer, Timm and Obersteiner, Michael},
+  year = {2010},
+  publisher = {Elsevier},
+  journal = {Energy Policy}
+}
+
+@article{Dietrich:2019,
+  doi = {https://doi.org/10.5194/gmd-12-1299-2019},
+  title = {MAgPIE 4 – a modular open-source framework for modeling global land systems},
+  author = {Dietrich, J. P. and Bodirsky, B. L. and Humpenöder, F. and Weindl, I. and Stevanović, M. and Karstens, K. and Kreidenweis, U. and Wang, X. and Mishra, A. and Klein, D. and Ambrósio, G. and Araujo, E. and Yalew, A. W. and Baumstark, L. and Wirth, S. and Giannousakis, A. and Beier, F. and Chen, D. M.-C. and Lotze-Campen, H. and Popp, A.},
+  year = {2019},
+  publisher = {Copernicus GmbH},
+  journal = {Geoscientific Model Development}
+}
+
+@article{Beach:2012,
+  doi = {https://doi.org/10.1142/S2010007812500017},
+  title = {Modeling Bioenergy, Land Use, And Ghg Emissions With Fasomghg: Model Overview And Analysis Of Storage Cost Implications},
+  author = {Beach, Robert H. and Zhang, Yuquan W. and Mccarl, Bruce A.},
+  year = {2012},
+  publisher = {World Scientific Publishing Co. Pte. Ltd.},
+  journal = {Climate Change Economics (CCE)}
+}
