+++ joss_paper/paper/paper.md
@@ -37,6 +37,8 @@ bibliography: paper.bib
 
 Ireland's agricultural landscape is dominated by grassland, supporting extensive dairy and suckler beef production. Investment in Ireland's bio-economy requires substantial changes to the AFOLU sector. `OptiGob` allows users to manipulate critical AFOLU levers that determine transition pathways to explore their impacts. `OptiGob` utilises data from GOBLIN Lite [@Duffy:2024; @DuffyB:2022], FERS-CBM [@Black:2025; @Kurz:2008], and LCAD2.0 [@Martinez-Acre:2025] to calculate outputs from agriculture, forestry, and anaerobic digestion. This allows users to explore land use and livestock trade-offs, leveraging the outputs from the upstream models without the overhead of running upstream modelling chains. To the authors' knowledge, `OptiGob` represents the first attempt to do so in the Irish context. 
 
+A range of integrated modelling tools exist to explore AFOLU interactions under climate and land constraints, including optimisation-based land-sector models (e.g., FASOM-GHG [@Beach:2012]) and regional or global partial-equilibrium frameworks such as CAPRI [@Leip:2008], GLOBIOM [@Havlík:2010], and MAgPIE [@Dietrich:2019]. These models typically operate at EU or global scales and solve equilibrium or dynamic land-allocation problems. `OptiGob` instead provides a lightweight, inventory-aligned optimisation layer tailored to rapid national policy exploration in Ireland.
+
 `OptiGob` estimates the net greenhouse gas (GHG) emissions from land use change and available emissions and land budget for grass-based livestock (Dairy and Suckler cow) production. `Pyomo` [@bynum2021pyomo; @hart2011pyomo] is used to optimise livestock populations, while respecting area commitment (afforestation, anaerobic digestion, BECCS (Bioenergy with Carbon Capture and Storage), protein crops) and emission (CO~2~e in the case of net-zero, or CH~4~ alongside net-zero CO~2~e for N~2~O and CO~2~ under a split gas target) constraints.
 
 The GOBLIN framework has been applied in recent studies of net-zero pathways for AFOLU [@Henn:2025; @Bishop:2024; @DuffyB:2022]. `OptiGob` builds upon this framework, providing a single-interface tool for exploring synergies and trade-offs across sectors.
