### 2026-02-20 #9271 @colmduff
> [@lionel42](https://github.com/lionel42) - Thank you for summarizing the remaining issues and for raising your concerns.
>
> [@colmduff](https://github.com/colmduff) - JOSS requires the paper to describe the current state of the field relevant to this software. As the reviewer suggested, please cite comparable tools and state their similarities and differences. It is important to explain the specific contributions and added value of this software compared to others. Please update the thread once you have addressed these comments. Thanks!

---
Thanks @mengqi-z

I have updated the Statement of Need section to cite comparable optimisation-based and land-system modelling tools (including FASOM-GHG, CAPRI, GLOBIOM, and MAgPIE) and to briefly clarify how OptiGob differs in scope and implementation..

Thanks to you, @quaz115 and @lionel42 for your time and careful consideration throughout the review process.
