### 2026-02-20 #9271 @mengqi-z
@colmduff - Thanks for your quick response! I think there is one other remaining issue raised by the reviewer:

> Also, the authors have claimed in a previous comment:
>> In the next update, we will add: A tutorial showing how to modify or replace the SQLite database to adapt the model to another region or dataset.
>
> I am not aware of such a tutorial on the current release.

Could you please respond to that?
