### 2026-02-20 #9271 @colmduff
> @colmduff - Thanks for your quick response! I think there is one other remaining issue raised by the reviewer:
>
> > Also, the authors have claimed in a previous comment:
> >> In the next update, we will add: A tutorial showing how to modify or replace the SQLite database to adapt the model to another region or dataset.
> >
> > I am not aware of such a tutorial on the current release.
>
> Could you please respond to that?

---
Apologies @mengqi-z, and thank you for raising this.

There was an earlier discussion regarding whether the SQLite database represented a single “use case.” I clarified at that time that the model is configured for a single regional context (Ireland), not a single analytical use case.

The SQLite database can technically be replaced. However, user-friendly support for modifying or substituting the database is still under active development. We wish to develop this to allow for users to utilize different regions or tool chains.

For that reason, a formal tutorial on adapting the model to other regions or datasets has not been included in the current release. Though, as stated, would be included in a future release.

As I understood it, the closure and deletion of the earlier issue had indicated that this point had been resolved in the context of the present version.

The current submission documents the released software as it stands.

Thanks again to you and the reviewers for the careful review
