### 2026-02-23 #9271 @lionel42
@colmduff Thanks for answering on that point.

Even if this solution is not optimal, it should be enough for users to know where to start their own case study.
It is an okay solution for the current state of this project and for the joss submission.

@mengqi-z  Thanks a lot for handling this review, I have no more comments for this review, you can continue with the process.
