### 2026-02-26 colmduff/OptiGob#30 @mengqi-z
JOSS Paper - Proofreading
Hi @colmduff,

Could you please address the following in `paper.md` and `paper.bib`?

1. On line 58, use inline code formatting for `Pyomo`, as you have done in other instances. I also suggest using inline code formatting for the names of all tools, packages, and models.
2. Please fix the DOIs in the following references in `paper.bib`. You can simply remove `https://doi.org/` from those DOI entries.
```
- https://doi.org/10.5194/bg-5-73-2008 is INVALID because of 'https://doi.org/' prefix
- https://doi.org/10.1016/j.enpol.2010.03.030 is INVALID because of 'https://doi.org/' prefix
- https://doi.org/10.5194/gmd-12-1299-2019 is INVALID because of 'https://doi.org/' prefix
- https://doi.org/10.1142/S2010007812500017 is INVALID because of 'https://doi.org/' prefix
```

For JOSS review openjournals/joss-reviews#9271
