### 2026-02-27 08923bc
updated paper.bib and paper.md
