+++ joss_paper/paper/paper.md
@@ -31,27 +31,27 @@ bibliography: paper.bib
 ---
 # Summary
 
-`OptiGob` is a Python-based tool designed to explore configurations of Ireland’s agriculture, forestry, and other land use (AFOLU) sectors. The main purpose is to assist users in the assessment of environmental and economic impact pathways, based on different land use transition pathways under varying assumptions about agriculture, afforestation, emissions abatement, and carbon dioxide removal (CDR) strategies. `OptiGob` combines data outputs from the GOBLIN Lite for agriculture and land use, from FERS-CBM for forestry, and from the LCAD2.0 for anaerobic digestion, to generate scenarios that respect biophysical constraints. `OptiGob` provides a flexible, customisable tool for researchers, policymakers, and educators to explore environmental and economic trade-offs associated with land use transition pathways.
+`OptiGob` is a Python-based tool designed to explore configurations of Ireland’s agriculture, forestry, and other land use (AFOLU) sectors. The main purpose is to assist users in the assessment of environmental and economic impact pathways, based on different land use transition pathways under varying assumptions about agriculture, afforestation, emissions abatement, and carbon dioxide removal (CDR) strategies. `OptiGob` combines data outputs from the `GOBLIN Lite` for agriculture and land use, from `FERS-CBM` for forestry, and from the `LCAD2.0` for anaerobic digestion, to generate scenarios that respect biophysical constraints. `OptiGob` provides a flexible, customisable tool for researchers, policymakers, and educators to explore environmental and economic trade-offs associated with land use transition pathways.
 
 # Statement of need
 
-Ireland's agricultural landscape is dominated by grassland, supporting extensive dairy and suckler beef production. Investment in Ireland's bio-economy requires substantial changes to the AFOLU sector. `OptiGob` allows users to manipulate critical AFOLU levers that determine transition pathways to explore their impacts. `OptiGob` utilises data from GOBLIN Lite [@Duffy:2024; @DuffyB:2022], FERS-CBM [@Black:2025; @Kurz:2008], and LCAD2.0 [@Martinez-Acre:2025] to calculate outputs from agriculture, forestry, and anaerobic digestion. This allows users to explore land use and livestock trade-offs, leveraging the outputs from the upstream models without the overhead of running upstream modelling chains. To the authors' knowledge, `OptiGob` represents the first attempt to do so in the Irish context. 
+Ireland's agricultural landscape is dominated by grassland, supporting extensive dairy and suckler beef production. Investment in Ireland's bio-economy requires substantial changes to the AFOLU sector. `OptiGob` allows users to manipulate critical AFOLU levers that determine transition pathways to explore their impacts. `OptiGob` utilises data from `GOBLIN Lite` [@Duffy:2024; @DuffyB:2022], `FERS-CBM` [@Black:2025; @Kurz:2008], and `LCAD2.0` [@Martinez-Acre:2025] to calculate outputs from agriculture, forestry, and anaerobic digestion. This allows users to explore land use and livestock trade-offs, leveraging the outputs from the upstream models without the overhead of running upstream modelling chains. To the authors' knowledge, `OptiGob` represents the first attempt to do so in the Irish context. 
 
[…]
