### 2026-02-27 #9271 @colmduff
Hi @mengqi-z ,

Not sure if the above is for me or for you, but if this helps:

1. PyPI: [https://pypi.org/project/optigob/ ](https://pypi.org/project/optigob/ )
2. Version: 0.1.5
3. Zenodo DOI: [10.5281/zenodo.17541475](10.5281/zenodo.17541475)
