### 2026-02-27 #9271 @colmduff
> > Hi [@mengqi-z](https://github.com/mengqi-z) ,
> > Not sure if the above is for me or for you, but if this helps:
> >
> > 1. PyPI: [https://pypi.org/project/optigob/ ](https://pypi.org/project/optigob/)
> > 2. Version: 0.1.5
> > 3. Zenodo DOI: [10.5281/zenodo.17541475](10.5281/zenodo.17541475)
>
> Thanks [@colmduff](https://github.com/colmduff). Usually I would ask you to release the latest version of the package and archive it for a DOI after fixing issues in the paper, but this works as well. Thanks for getting these done.
>
> I noticed that the license on the Zenodo archive doesn’t match the one on GitHub/PyPI. Could you please update the license in the Zenodo archive to match the one used in your GitHub repository (MIT)?

---
thanks @mengqi-z , I have updated that now.
