### 2026-02-27 #9271 @mengqi-z
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/22501249712?check_suite_focus=true).

@openjournals/dev Could you please look into this error and advise on how to resolve it? Thank you.
