### 2026-03-02 #9271 @mengqi-z
@colmduff – I’ve recommended your submission for acceptance. Thanks so much for submitting to JOSS! From here, our Editor-in-Chief will step in to do some final checks before the paper can be officially accepted and published. Congrats and great job! 🎉

@quaz115, @lionel42 – Huge thanks to both of you for reviewing this paper! It’s been great working with you during this process. I look forward to seeing your reviews for JOSS in the future!
