### 2026-03-06 #9271 @kthyng
@colmduff

* can you make a release that matches the version associated with this submission?

paper:
- **Page 7 – Reference (Leip et al., 2008)**
  Proper nouns not capitalized.
  - Current: `european agriculture` / `europe`
  - Suggested: `European agriculture` / `Europe`

- **Page 7 – Reference (Hart et al., 2011)**
  Programming language name not capitalized.
  - Current: `optimization modeling in python`
  - Suggested: `optimization modeling in Python`
