### 2026-03-07 #9271 @colmduff
> [@colmduff](https://github.com/colmduff)
>
> * can you make a release that matches the version associated with this submission?
>
> paper:
>
> * **Page 7 – Reference (Leip et al., 2008)**
>   Proper nouns not capitalized.
>
>   * Current: `european agriculture` / `europe`
>   * Suggested: `European agriculture` / `Europe`
> * **Page 7 – Reference (Hart et al., 2011)**
>   Programming language name not capitalized.
>
>   * Current: `optimization modeling in python`
>   * Suggested: `optimization modeling in Python`

---
Hi @kthyng

Apologies, I might be missing something, but the current version is 0.1.5 (incorporating the changes requested to the code base).

That release is on [PyPI](https://pypi.org/project/optigob/) and [Zenodo](https://doi.org/10.5281/zenodo.18799219)

Happy to update if I have missed something?

The two paper edits also seem to have already been addressed in the bib as well, so I am not sure why the system is flagging that?
