### 2026-03-07 #9271 @colmduff
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.09271/joss.09271/10.21105.joss.09271.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.09271/joss.09271/10.21105.joss.09271.pdf) 📄 👈

---
@kthyng , I did a quick sanity check here, on the capitalisation issue here. Both are definitely changed in the repo, but that is not being transalted in the proof. So, I am not sure what is going on there, or if the issue is on my end?
