### 2026-03-17 #9271 @kthyng
@colmduff

* Sorry I was referring to a release on GitHub which people usually also do. It is helpful for tracking the releases but not mandatory if you're doing it elsewhere.
