### 2026-03-17 #9271 @kthyng
@colmduff I believe you haven't added enough {} in your bib entry. One set of {} around the title does not force capitalization: Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
