+++ joss_paper/paper/paper.bib
@@ -63,7 +63,7 @@
 @article{Henn:2025,
   doi = {10.1016/j.scitotenv.2025.179115},
   author = {Henn, D. and Duffy, C. and Styles, D.},
-  title = {Exploring the frontier of bovine protein production within territorial net zero emission targets},
+  title = {{Exploring the frontier of bovine protein production within territorial net zero emission targets}},
   journal = {Science of The Total Environment},
   year = 2025
 }
@@ -71,7 +71,7 @@
 @article{Costa:2023,
   doi = {10.1016/j.jclepro.2023.138826},
   author = {Costa, M.P. and Saget, S. and Zimmermann, B. and Petig, E. and Angenendt, E. and Rees, R.M. and Chadwick, D. and Gibbons, J. and Shrestha, S. and Williams, M. and Styles, D.},
-  title = {Environmental and land use consequences of replacing milk and beef with plant-based alternatives},
+  title = {{Environmental and land use consequences of replacing milk and beef with plant-based alternatives}},
   journal = {Journal of Cleaner Production},
   year = 2023
 }
@@ -79,17 +79,9 @@
 @article{Styles:2022,
   doi = {10.1016/j.jclepro.2022.130441},
   author = {Styles, D. and Yesufu, J. and Bowman, M. and Williams, A.P. and Duffy, C. and Luyckx, K.},
-  title = {Climate mitigation efficacy of anaerobic digestion in a decarbonising economy},
+  title = {{Climate mitigation efficacy of anaerobic digestion in a decarbonising economy}},
   journal = {Journal of Cleaner Production},
   year = 2022
-}
-
-@article{McEniry:2013,
-  author = {McEniry, J. and Crosson, P. and Finneran, E. and Mcgee, M. and Keady, T.W.J. and O’Kiely, P.},
-  title = {How much grassland biomass is available in {Ireland} in excess of livestock requirements?},
-  journal = {Irish Journal of Agricultural and Food Research},
-  year = {2013},
-  url ={http://www.jstor.org/stable/23631018}
 }
 
 @misc{Github,
@@ -102,7 +94,7 @@
 
 @book{bynum2021pyomo,
   doi = {10.1007/978-3-030-68928-5},
-  title={Pyomo--optimization modeling in python}, 
+  title={{Pyomo--optimization modeling in Python}}, 
   author={Bynum, Michael L. and Hackebeil, Gabriel A. and Hart, William E. and Laird, Carl D. and Nicholson, Bethany L. and Siirola, John D. and Watson, Jean-Paul and Woodruff, David L.}, 
   edition={Third}, 
   volume={67}, 
@@ -111,7 +103,7 @@
 
 @article{hart2011pyomo, 
     doi = {10.1007/s12532-011-0026-8},
-    title={Pyomo: modeling and solving mathematical programs in Python}, 
+    title={{Pyomo: modeling and solving mathematical programs in Python}}, 
     author={Hart, William E and Watson, Jean-Paul and Woodruff, David L}, 
     journal={Mathematical Programming Computation}, 
     volume={3}, 
@@ -132,7 +124,7 @@
 
 @article{Leip:2008,
   doi = {10.5194/bg-5-73-2008},
-  title = {Linking an economic model for European agriculture with a mechanistic model to estimate nitrogen and carbon losses from arable soils in Europe},
+  title = {{Linking an economic model for European agriculture with a mechanistic model to estimate nitrogen and carbon losses from arable soils in Europe}},
   author = {Leip, A. and Marchi, G. and Koeble, R. and Kempen, M. and Britz, W. and Li, C.},
[…]
