### 2026-03-19 #9271 @colmduff
> [@colmduff](https://github.com/colmduff) I believe you haven't added enough {} in your bib entry. One set of {} around the title does not force capitalization: Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.

---
Thanks @kthyng, I have updated the .bib file now.
