### 2026-03-19 #9271 @colmduff
> [@colmduff](https://github.com/colmduff)
>
> * Sorry I was referring to a release on GitHub which people usually also do. It is helpful for tracking the releases but not mandatory if you're doing it elsewhere.

---
@kthyng , oh sorry, I have done that now.
